pynet/requirements.txt
async-timeout>=5.0.1
asyncio>=3.4.3
asyncpg>=0.30.0
contourpy>=1.3.1
cycler>=0.12.1
fonttools>=4.55.3
kiwisolver>=1.4.8
matplotlib>=3.10.0
natsort>=8.4.0
networkx>=3.4.2
numpy>=2.2.1
packaging>=24.2
pandas>=2.2.3
pillow>=11.1.0
psycopg2>=2.9.10
psycopg2-binary>=2.9.10
pyparsing>=3.2.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2024.2
scapy>=2.6.1
scipy>=1.15.1
six>=1.17.0
tzdata>=2024.2
